Adding support for filtering whatif json output (#482)
* Adding support for filtering whatif results using the ExcludeChangeType parameter for the Get-Az****DeploymentWhatIfResult cmdlets. This is in response to issue 475 that I opened. I propose using a PSF config value for controlling this i.e. AzOps.Core.WhatifExcludedChangeTypes which would be an array containing the list of change types to exclude. https://docs.microsoft.com/en-us/azure/azure-resource-manager/templates/deploy-what-if?tabs=azure-powershell#change-types

Also I am suggesting a slight change to the formatting of output.json such that the output will be a valid JSON document when there is more than 1 template deployed with changes. Currently the file has a simple text append function which when processing multiple templates in a single run will produce a ill formatted document.